pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "political-ad-collector"
version = "0.1.0"
description = "Modular framework for collecting political advertising data from multiple platforms"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
    "tenacity>=8.2.3",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]
gcp = [
    "google-cloud-storage>=2.14.0",
    "google-cloud-secret-manager>=2.16.4",
    "google-cloud-bigquery>=3.14.0",
    "functions-framework>=3.0.0",
]

[project.scripts]
collect-meta-ads = "scripts.run_meta_collector:main"
test-ad-credentials = "scripts.test_credentials:main"
upload-to-bigquery = "scripts.upload_to_bigquery:main"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | output
  | \.checkpoints
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["output/*", ".checkpoints/*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "output/",
    ".checkpoints/",
    "tests/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=collectors --cov=shared --cov-report=term-missing"

[tool.coverage.run]
source = ["collectors", "shared"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
